# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /sparks/earn:
    post:
      operationId: claimDailyEarn
      summary: Claim a daily Sparks earn action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EarnClaimInput"
      responses:
        "200":
          description: Updated balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SparksBalance"
        "409":
          description: Already claimed today
  /sparks/purchase:
    post:
      operationId: purchaseSparks
      summary: Simulate purchasing a Sparks bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "200":
          description: Updated balance after purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SparksBalance"
components:
  schemas:
...
    SparksBalance:
      type: object
      required: [balance, earned_this_week, spent_this_week]
      properties:
        balance:
          type: integer
        earned_this_week:
          type: integer
        spent_this_week:
          type: integer
        free_keys_remaining:
          type: integer
        free_unsends_remaining:
          type: integer
        daily_earn_available:
          type: array
          items:
            $ref: "#/components/schemas/DailyEarnOption"
    DailyEarnOption:
      type: object
      required: [type, amount, claimed, label]
      properties:
        type:
          type: string
        label:
          type: string
        amount:
          type: integer
        claimed:
          type: boolean
...
    EarnClaimInput:
      type: object
      required: [claim_type]
      properties:
        claim_type:
          type: string
    PurchaseInput:
      type: object
      required: [bundle_id]
      properties:
        bundle_id:
          type: string