# lib/api-zod/src/generated/types/dailyEarnOption.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface DailyEarnOption {
  type: string;
  label: string;
  amount: number;
  claimed: boolean;
}