# lib/api-zod/src/generated/types/sparksBalance.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { DailyEarnOption } from './dailyEarnOption';

export interface SparksBalance {
  balance: number;
  earned_this_week: number;
  spent_this_week: number;
  free_keys_remaining?: number;
  free_unsends_remaining?: number;
  daily_earn_available?: DailyEarnOption[];
}